move side effects out of assembler
