+++ aiuser/consent/__init__.py
@@ -1,3 +1,4 @@
-from aiuser.consent.service import CONSENT_EMBED_TITLE, ConsentService
+from aiuser.consent.service import ConsentService
+from aiuser.consent.view import CONSENT_EMBED_TITLE, maybe_send_consent_embed
 
-__all__ = ["ConsentService", "CONSENT_EMBED_TITLE"]
+__all__ = ["ConsentService", "CONSENT_EMBED_TITLE", "maybe_send_consent_embed"]
